+++ Code/src/ConferenceGraph.cpp
@@ -9,22 +9,22 @@ Graph<int> ConferenceGraph::buildGraph() const {
     graph.addVertex(getSinkID());
 
     //Add vertices for submissions and connect them to the source vertex
-    for (int i = 0; i < data.submissions.size(); i++) {
+    for (size_t i = 0; i < data.submissions.size(); i++) {
         int subID = getSubmissionID(i);
         graph.addVertex(subID);
         graph.addEdge(getSourceID(), subID, data.params.minReviewsPerSubmission);
     }
 
     //Add vertices for reviewers and connect them to the sink vertex ()
-    for (int i = 0; i < data.reviewers.size(); i++) {
+    for (size_t i = 0; i < data.reviewers.size(); i++) {
         int revID = getReviewerID(i);
         graph.addVertex(revID);
         graph.addEdge(revID, getSinkID(), data.params.maxReviewsPerReviewer);    
     }
 
     //Connect submission vertices to reviewer vertices based on domain matching
-    for(int i = 0; i < data.submissions.size(); i++) {
-        for (int j = 0; j < data.reviewers.size(); j++) {
+    for(size_t i = 0; i < data.submissions.size(); i++) {
+        for (size_t j = 0; j < data.reviewers.size(); j++) {
             if (domainsMatch(data.reviewers[j], data.submissions[i])) {
                 graph.addEdge(getSubmissionID(i), getReviewerID(j), 1); //Connect submission vertices to reviewer vertices with capacity 1
             }
@@ -35,12 +35,12 @@ Graph<int> ConferenceGraph::buildGraph() const {
 }
 
 
-int ConferenceGraph::getReviewerID(int index) const {
-    return index + 1 + data.submissions.size();
+int ConferenceGraph::getReviewerID(size_t index) const {
+    return static_cast<int>(index + 1 + data.submissions.size());
 }
 
-int ConferenceGraph::getSubmissionID(int index) const {
-    return index + 1;
+int ConferenceGraph::getSubmissionID(size_t index) const {
+    return static_cast<int>(index + 1);
 }
 
 int ConferenceGraph::getSourceID() const {
@@ -48,7 +48,7 @@ int ConferenceGraph::getSourceID() const {
 }
 
 int ConferenceGraph::getSinkID() const {
-    return data.reviewers.size() + data.submissions.size() + 1;
+    return static_cast<int>(data.reviewers.size() + data.submissions.size() + 1);
 }
 
 bool ConferenceGraph::domainsMatch(const Reviewer& rev, const Submission& sub) const {
